// src/app/api/generated/model/ProcessTransactionContext.ts
/**
 * Wari teller API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface ProcessTransactionContext {
    securityTokenTransaction?: boolean;

    tokenValidationDesc?: string;

    wariPassOTP?: string;

    wariPassToken?: string;

}